работа_с_bigartm/requirements.txt
# зависимости для получения наиболее релевантных тем для нового текста (из pdf):
bigartm==0.9.2
nltk==3.8.1
gensim==4.3.1
pymystem3==0.2.0

